Rewrite createStore Rewrite createStore to use snapshots. This should make it more generally useful, rather than specific to the task stories. Update MockTask to initialise the store from a snapshot.
